feat: wip hugo site - not ready
Lots of experimentations with trying to create a default theme.  There's a
lot of work here that's worth saving with regards to rewriting the content
in AsciiDoc, but the theme likely should start from scratch.

I wanted to get my hands on things and push and pull every lever that was
available to see how things worked.  Next I'll read through the docs top
to bottom and give it a second try.